+++ tests/testthat/test-nest.R
@@ -39,15 +39,27 @@ test_that("nesting everything yields a simple data frame", {
   expect_equal(out$data[[1L]], df)
 })
 
+test_that("nest preserves order of data", {
+  df <- tibble(x = c(1, 3, 2, 3, 2), y = 1:5)
+  out <- nest(df, y)
+  expect_equal(out$x, c(1, 3, 2))
+})
+
+test_that("empty factor levels don't affect nest", {
+  df <- tibble(
+    x = factor(c("z", "a"), levels = letters),
+    y = 1:2
+  )
+  out <- nest(df, y)
+  expect_equal(out$x, df$x)
+})
+
 test_that("nesting works for empty data frames", {
   df <- tibble(x = 1:3, y = c("B", "A", "A"))[0, ]
 
   out <- nest(df, x)
   expect_equal(names(out), c("y", "data"))
 
-  if (utils::packageVersion("dplyr") > "0.7.99") {
-    expect_equal(unnest(out), df)
-  }
   expect_equal(nrow(out), 0L)
   expect_equal(length(out$data), 0L)