+++ NEWS.md
@@ -1,10 +1,31 @@
-# tidyr 0.8.2.9000
+# tidyr (development version)
+
+* `complete()` now uses `full_join()` so that all levels are preserved even
+  when not all levels are specified (@Ryo-N7, #493).
+
+* `gather()` throws an error if a column is a data frame (#553)
+
+* New `expand_grid()` function which is a tidy version of `expand.grid()`.
+  It is lower-level than the existin `expand()` and `crossing()` functions as
+  it takes individual vectors, and does not sort or uniquify them.
+
+* `crossing()`, `nesting()`, and `expand()` have been rewritten to use 
+  the vctrs package. This should not affect much existing code, but
+  considerably simplies the implementation and ensures that these functions 
+  work consistently across all generalised vectors (#557). As part of this
+  alignment, these functions now only drop `NULL` inputs, not any 0-length
+  vector.
+
+* `crossing()` now takes the unique values of data frame inputs, not just
+  vector inputs (#490).
+
+# tidyr 0.8.3
 
 * `crossing()` preserves factor levels (#410), now works with list-columns 
   (#446, @SamanthaToet). (These also help `expand()` which is built on top
   of `crossing()`)
 
-* `nest()` uses `dplyr::group_nest()` when available. 
+* `nest()` is compatible with dplyr 0.8.0.
 
 * `spread()` works when the id variable has names (#525).